requirements.txt
# Core dependencies
torch>=1.10.0
transformers>=4.15.0
diffusers>=0.10.0
accelerate>=0.12.0
Pillow>=9.0.0
numpy>=1.20.0

# API integrations
openai>=1.0.0
anthropic>=0.5.0
replicate>=0.15.0
huggingface-hub>=0.12.0

# Server dependencies
fastapi>=0.70.0
uvicorn>=0.15.0
websockets>=10.0
python-multipart>=0.0.5

# Utilities
python-dotenv>=0.19.0
pydantic>=1.8.0
tqdm>=4.62.0
loguru>=0.6.0
typer>=0.4.0
rich>=10.0.0

# Storage and caching
diskcache>=5.2.0
sqlalchemy>=1.4.0

# Image processing
opencv-python>=4.5.0
scikit-image>=0.19.0

# Scientific and optimization
scipy>=1.7.0
scikit-learn>=1.0.0
sympy>=1.9.0

# Development tools
pytest>=6.0.0
black>=21.5b2
isort>=5.8.0
mypy>=0.812
flake8>=4.0.0
